lab 10 - intermediary

+++ lab-10/nft-staking/programs/nft-staking/src/state/stake_config.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct StakeConfig {
     pub authority: Pubkey,
     pub max_stake: u8,
-    pub freeze_period: i32,
+    pub freeze_period: i64,
     pub points_per_stake: u8,
     pub rewards_bump: u8,
     pub bump: u8,
